领养中心，图片查看，更多资料，签到。。。。

+++ CommunityProject/CommunityProject/Chat/C/DiscoverViewController.m
@@ -127,7 +127,7 @@
                 if (arr.count != 0) {
                     weakSelf.unreadArr = data[@"data"];
                     weakSelf.msgLabel.hidden = NO;
-                    weakSelf.msgLabel.text = [NSString stringWithFormat:@"%ld",arr.count];
+                    weakSelf.msgLabel.text = [NSString stringWithFormat:@"%ld",(unsigned long)arr.count];
                     int i = 0;
                     for (NSDictionary * dic in arr) {
                         if (i == 0) {
